Restrict login during backup if safe to do so.
If the root account is accessible, ('P' from passwd -S), then deny login access to non-root users for the duration of the backup. If root is not accessible, then do not deny access as it would be dangerous if for some reason it is not re-allowed. Only root would have access. (Even though shutdown clears the restriction, this script does not demand a shutdown).