Use absolute bash path in script shebangs
Pass the absolute bash path from configure to Makefiles to use in the
shebang of script files. This satisfies some rpm build tools and the
shebang can pass only 1 argument so this is the most flexible.
e.g.
This does not work as argument 1 is to env and is seen as the complete
awk -f whereas -f should be the argument to awk.
So
This does work as -f is the one argument and is passed to awk.

Using env is OK for non-AutoTools scripts.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>