logrotate file: create - remove owner and perm
netbckup creates the logfile with correct ownership and permissions.
logrotate does nothing if the log file does not exist ie before netbckup
is run for the first time. Subsequently a plain create will use
ownership and permissions from the file being rotated. So, remove the
ownership and permissions on create.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>