Prepend own exec invocations with sbindir
If an AutoTool and package installation of the package exists then the
executable invoked by a script will depend on the ordering of the PATH
variable. To ensure that AT and package versions work independently
prepend the executable name with sbindir making the packages
self-contained even when both are installed.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>